+++ test/sample_data/get_user_by_id.py
@@ -1,0 +1,25 @@
+from datetime import datetime
+from dateutil.tz import tzutc
+
+{
+    'id': 1,
+    'login': 'kueda',
+    'spam': False,
+    'suspended': False,
+    'created_at': datetime(2008, 3, 20, 21, 15, 42, tzinfo=tzutc()),
+    'login_autocomplete': 'kueda',
+    'login_exact': 'kueda',
+    'name': 'Ken-ichi Ueda',
+    'name_autocomplete': 'Ken-ichi Ueda',
+    'orcid': 'https://orcid.org/0000-0003-0145-6846',
+    'icon': 'https://static.inaturalist.org/attachments/users/icons/1/thumb.jpg?1475527316',
+    'observations_count': 40947,
+    'identifications_count': 95624,
+    'journal_posts_count': 86,
+    'activity_count': 136657,
+    'species_count': 8831,
+    'universal_search_rank': 40947,
+    'roles': ['admin', 'curator'],
+    'site_id': 1,
+    'icon_url': 'https://static.inaturalist.org/attachments/users/icons/1/medium.jpg?1475527316',
+}